move assets and navebar folders to out of app folder
